+++ packages/entities/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./entity-ref";
 export * from "./uri";
+export * from "./bech32";